requirements.txt
# 核心爬虫库
requests==2.31.0
beautifulsoup4==4.12.2
scrapy==2.11.0
selenium==4.15.2

# 数据处理
pandas==2.1.3
numpy==1.26.4
openpyxl==3.1.2
xlrd==2.0.1

# 数据库
sqlalchemy==2.0.23
pymongo==4.6.0
redis==5.0.1
pymysql==1.1.0

# 异步支持
aiohttp==3.9.1
asyncio-throttle==1.0.2

# 配置和日志
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
pydantic-settings==2.1.0
PyYAML==6.0.1

# 工具库
lxml==4.9.3
cssselect==1.2.0
fake-useragent==1.4.0
python-dateutil==2.8.2
tqdm==4.66.1

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

